Changed Ratings and Changed avgRating Type
Changed Ratings for a few sample movie reviews and movie entries.  Also
changed avgRating type in Movie Table from int to decimal(2,1).

+++ Create Database/Create Database and Tables.sql
@@ -1,5 +1,5 @@
 /* CREATE THE ENTIRE YellowBucketCSC365 DATABASE, AND CREATE/LINK ALL TABLES */ /* Austin Caldwell and Evan Wehr */
--- Updated 10/21/2015
+-- Updated 10/22/2015
 
 -- Create the database itself
 Use master
@@ -21,7 +21,7 @@ CREATE TABLE Movie -- Create a table of movies
 	movieID int PRIMARY KEY IDENTITY(0000,1) NOT NULL,
 	title varchar(64) NOT NULL,
 	movieDescription varchar(256) NOT NULL,
-	avgRating int CONSTRAINT avgRating_Default DEFAULT 0 CONSTRAINT avgRating_CC CHECK(avgRating >= 0 AND avgRating <=5) NOT NULL,
+	avgRating decimal(2,1) CONSTRAINT avgRating_Default DEFAULT 0 CONSTRAINT avgRating_CC CHECK(avgRating >= 0 AND avgRating <=5) NOT NULL,
 	director varchar(32) NOT NULL,
 	studio varchar(32) NOT NULL,
 	runTime int CONSTRAINT runTime_CC Check(runTime >= 0) NOT NULL,
